+++ app/datatables/customer_datatable.rb
@@ -24,8 +24,8 @@ class CustomerDatatable < AjaxDatatablesRails::Base
         record.account_number,
         record.phone_one,
         link_to("", edit_customer_path(record), class: "glyphicon glyphicon-pencil"),
-        link_to("", customer_path(record), class: "glyphicon glyphicon-trash",  method: :delete),
-        link_to("Job", create_job_path(record), class: "glyphicon glyphicon-new-window",  method: :post),
+        link_to("", customer_path(record), data:{confirm: 'Are you sure?'}, class: "glyphicon glyphicon-trash",  method: :delete),
+        link_to("Job", create_job_path(record), data:{confirm: 'Are you sure?'}, class: "glyphicon glyphicon-new-window",  method: :post),
         link_to("Jobs", customer_job_path(record), class: "glyphicon glyphicon-search")
 
         # comma separated list of the values for each cell of a table row
